Requested changes, static analysis tests

+++ tests/fixture/Point.php
@@ -7,8 +7,8 @@ namespace Psl\Tests\Fixture;
 final class Point
 {
     public function __construct(
-        public readonly float $x,
-        public readonly float $y,
+        public readonly int $x,
+        public readonly int $y,
     ) {
     }
 }
